+++ tests/Unit/AltchaBundleTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Huluti\AltchaBundle\Tests\Unit;
+namespace Tito10047\AltchaBundle\Tests\Unit;
 
-use Huluti\AltchaBundle\DependencyInjection\Compiler\HulutiAltchaBundleCompilerPass;
-use Huluti\AltchaBundle\HulutiAltchaBundle;
+use Tito10047\AltchaBundle\DependencyInjection\Compiler\AltchaBundleCompilerPass;
+use Tito10047\AltchaBundle\AltchaBundle;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
@@ -14,13 +14,13 @@ class AltchaBundleTest extends TestCase
     public function testBuild(): void
     {
         $container = new ContainerBuilder();
-        $bundle = new HulutiAltchaBundle();
+        $bundle = new AltchaBundle();
         $bundle->build($container);
 
         $this->assertNotEmpty(array_filter(
             $container->getCompilerPassConfig()->getPasses(),
             function ($value) {
-                return $value instanceof HulutiAltchaBundleCompilerPass;
+                return $value instanceof AltchaBundleCompilerPass;
             }
         ));
     }